+++ generators/typescript-v2/ast/src/index.ts
@@ -2,6 +2,7 @@ export * from "./ast/core/index.js";
 export {
     NamingConfigSchema,
     NamingObjectSchema,
+    resolveNoSerdeLayer,
     TypescriptCustomConfigSchema
 } from "./custom-config/TypescriptCustomConfigSchema.js";
 export * as ts from "./typescript.js";