+++ seed/csharp-sdk/inferred-auth-implicit/src/SeedInferredAuthImplicit/Simple/ISimpleClient.cs
@@ -2,7 +2,7 @@ namespace SeedInferredAuthImplicit;
 
 public partial interface ISimpleClient
 {
-    Task GetSomethingAsync(
+    WithRawResponseTask GetSomethingAsync(
         RequestOptions? options = null,
         CancellationToken cancellationToken = default
     );