+++ seed/csharp-sdk/oauth-client-credentials-mandatory-auth/unified-client-options/src/SeedOauthClientCredentialsMandatoryAuth/Simple/ISimpleClient.cs
@@ -2,7 +2,7 @@ namespace SeedOauthClientCredentialsMandatoryAuth;
 
 public partial interface ISimpleClient
 {
-    Task GetSomethingAsync(
+    WithRawResponseTask GetSomethingAsync(
         RequestOptions? options = null,
         CancellationToken cancellationToken = default
     );