+++ seed/csharp-sdk/unions/no-discriminated-unions/src/SeedUnions/Bigunion/BigunionClient.cs
@@ -29,7 +29,7 @@ public partial class BigunionClient : IBigunionClient
                 new JsonRequest
                 {
                     Method = HttpMethod.Get,
-                    Path = string.Format("/{0}", ValueConvert.ToPathParameterString(id)),
+                    Path = string.Format("/bigunion/{0}", ValueConvert.ToPathParameterString(id)),
                     Headers = _headers,
                     Options = options,
                 },
@@ -106,7 +106,7 @@ public partial class BigunionClient : IBigunionClient
                 new JsonRequest
                 {
                     Method = HttpMethodExtensions.Patch,
-                    Path = "",
+                    Path = "/bigunion",
                     Body = request,
                     Headers = _headers,
                     Options = options,
@@ -184,7 +184,7 @@ public partial class BigunionClient : IBigunionClient
                 new JsonRequest
                 {
                     Method = HttpMethodExtensions.Patch,
-                    Path = "/many",
+                    Path = "/bigunion/many",
                     Body = request,
                     Headers = _headers,
                     Options = options,